--- src/system/utils/index.ts
@@ -1,3 +1,0 @@
-export * from './actor';
-export * from './generic';
-export * from './templates';
